+++ vibratospace/src/matlab/sinusoidal_modeling.m~
@@ -1,0 +1,24 @@
+% TODO: Refactor to inherit these from Python (when calling script?).
+PICKLE_PATH = '/Users/maxsolomonhenry/Documents/Python/vibrato_space/vibratospace/data/data.pickle';
+SM_PATH = '/Users/maxsolomonhenry/Documents/MATLAB/sinusoidal-model';
+PITCH_RATE = 100;
+
+% Load SM library.
+addpath(genpath(SM_PATH));
+
+% Load pickle.
+fid = py.open(PICKLE_PATH, 'rb');
+data = py.pickle.load(fid);
+
+% Set up sinusoidal model.
+run('set_parameters.m');
+
+
+% Step through pitch trajectories.
+for k = 1:length(data)
+	temp = data{k}{'frequency'};
+	plot(temp);
+end
+
+[amplitude,frequency,phase,nsample,ds,center_frame,npartial,nframe] = ...
+    sinusoidal_analysis(wav,framelen,hop,nfft,fs,maxnpeak,relthres,absthres,delta,winflag,cfwflag{cf},normflag,zphflag,magflag{mf},ptrackflag{ptrck});
