+++ vibratospace/src/python/defaults.py
@@ -12,16 +12,16 @@ PITCH_RATE = 100
 EPS = 1e-6
 
 # Pickle filename for saving data.
-PICKLE_PATH = '../data/data.pickle'
+PICKLE_PATH = '../../data/data.pickle'
 
 # Path for processed audio.
-PROCESSED_PATH = '../audio/processed/'
+PROCESSED_PATH = '../../audio/processed/'
 
 # Path for raw audio (to be analyzed, re-synthesized).
 RAW_AUDIO_PATH = '../audio/raw/**/*.wav'
 
 # Path for figures.
-FIG_PATH = '../figs/'
+FIG_PATH = '../../figs/'
 
 # Path for audio test outputs.
-AUDIO_TEST_PATH = '../audio/tests/'
+AUDIO_TEST_PATH = '../../audio/tests/'
