+++ vibratospace/src/python/oscillators.py
@@ -1,6 +1,6 @@
 import numpy as np
-from vibratospace.src.defaults import SAMPLE_RATE, EPS
-from vibratospace.src.util import normalize, fix_length
+from vibratospace.src.python.defaults import SAMPLE_RATE, EPS
+from vibratospace.src.python.util import normalize, fix_length
 
 
 class Blit:
